refactor: densify six app-test allowlist suites under 400 lines
Continue #4352 Slice D by extracting shared fixtures and concern-splitting
CtSpacing, CtToggleSwitch, diplomacy detail, overture overlay, and naval
panel helpers so the shrink-only allowlist can drop those six entries.
