+++ app/test/diplomacy_detail_screen_test.dart
@@ -14,141 +14,10 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/misc.dart' show Override;
 import 'package:flutter_test/flutter_test.dart';
 
-import 'app_shell_harness.dart';
 import 'widget_test_assets.dart';
-
-const _humanId = 'gp1';
-const _otherId = 'gp2';
-
-Game _minimalGame({
-  DiplomaticEventType eventType = DiplomaticEventType.peace,
-  bool includeHistory = false,
-  bool includeDossier = false,
-  bool atWar = false,
-  int score = 70,
-  bool formalAlliance = false,
-}) {
-  final relation = DiplomacyRelation(
-    factionId1: _humanId,
-    factionId2: _otherId,
-    score: score,
-    state: atWar ? RelationState.atWar : RelationState.atPeace,
-    formalAlliance: formalAlliance,
-  );
-  return Game(
-    id: 'test',
-    worldState: WorldState(
-      turnState: TurnState(phase: TurnPhase.orders, turnNumber: 1),
-      oldWorld: const RegionData(),
-      newWorld: const RegionData(),
-    ),
-    turnTimeMapping: TurnTimeMapping.gdd01,
-    players: [
-      Player(id: _humanId, displayName: 'Human GP', isHuman: true, treasury: 0),
-      Player(
-        id: _otherId,
-        displayName: 'Other GP',
-        isHuman: false,
-        treasury: 0,
-      ),
-    ],
-    diplomacyRelations: [relation],
-    diplomaticHistoryEvents: includeHistory
-        ? [
-            DiplomaticEvent(
-              turn: 2,
-              intraTurnIndex: 0,
-              type: eventType,
-              participants: {_humanId, _otherId},
-              fromFactionId: _humanId,
-              toFactionId: _otherId,
-            ),
-          ]
-        : const [],
-    dossierEvidenceEntries: includeDossier
-        ? [
-            DossierEvidenceEntry(
-              observerId: _humanId,
-              subjectId: _otherId,
-              agendaType: 'test_agenda',
-              turnNumber: 3,
-              description: 'evidence-1',
-            ),
-          ]
-        : const [],
-  );
-}
-
-Future<void> _pumpDetail(
-  WidgetTester tester, {
-  required Game game,
-  required String factionId,
-  required String factionDisplayName,
-  required FactionKind kind,
-  DiplomacyRelation? relation,
-  List<Override> overrides = const <Override>[],
-}) {
-  return pumpAppShell(
-    tester,
-    overrides: overrides,
-    child: DiplomacyDetailScreen(
-      game: game,
-      humanPlayerId: _humanId,
-      factionId: factionId,
-      factionDisplayName: factionDisplayName,
-      kind: kind,
-      relation: relation,
-    ),
-    settle: true,
-  );
-}
-
-Future<void> _pumpOtherGp(
-  WidgetTester tester, {
-  required Game game,
-  DiplomacyRelation? relation,
-  FactionKind kind = FactionKind.greatPower,
-  List<Override> overrides = const <Override>[],
-}) {
-  return _pumpDetail(
-    tester,
-    game: game,
-    factionId: _otherId,
-    factionDisplayName: 'Other GP',
-    kind: kind,
-    relation: relation,
-    overrides: overrides,
-  );
-}
-
-String _displayNameFor(Game game, String id) {
-  final p = game.playerById(id);
-  if (p != null) return p.displayName;
-  for (final m in game.minorNations) {
-    if (m.id == id) return m.displayName ?? m.id;
-  }
-  for (final t in game.tribes) {
-    if (t.id == id) return t.displayName ?? t.id;
-  }
-  return id;
-}
-
-FactionKind _kindFor(Game game, String id) {
-  if (game.players.any((p) => p.id == id)) return FactionKind.greatPower;
-  if (game.minorNations.any((m) => m.id == id)) return FactionKind.minor;
-  return FactionKind.tribe;
-}
-
-List<String> _otherFactionIds(Game game) => <String>[
-  ...game.players.map((p) => p.id),
-  ...game.minorNations.map((m) => m.id),
-  ...game.tribes.map((t) => t.id),
-].where((id) => id != _humanId).toList();
-
-typedef _Pin = (Finder finder, Matcher matcher);
+import 'diplomacy_detail_screen_test_support.dart';
 
 void main() {
   suppressLogsForTests();
@@ -160,24 +29,35 @@ void main() {
   testWidgets(
     'DiplomacyDetailScreen renders either empty or non-empty history',
     (WidgetTester tester) async {
-      final game = _minimalGame();
-      final allFactionIds = _otherFactionIds(game);
+      final game = diplomacyDetailMinimalGame();
+      final allFactionIds = diplomacyDetailOtherFactionIds(game);
       String? nonEmptyHistoryFactionId;
       for (final id in allFactionIds) {
-        if (diplomaticHistoryForPair(game, _humanId, id).isNotEmpty) {
+        if (diplomaticHistoryForPair(
+          game,
+          diplomacyDetailHumanId,
+          id,
+        ).isNotEmpty) {
           nonEmptyHistoryFactionId = id;
           break;
         }
       }
       final chosenFactionId = nonEmptyHistoryFactionId ?? allFactionIds.first;
-      final history = diplomaticHistoryForPair(game, _humanId, chosenFactionId);
-      await _pumpDetail(
+      final history = diplomaticHistoryForPair(
+        game,
+        diplomacyDetailHumanId,
+        chosenFactionId,
+      );
+      await pumpDiplomacyDetail(
         tester,
         game: game,
         factionId: chosenFactionId,
-        factionDisplayName: _displayNameFor(game, chosenFactionId),
-        kind: _kindFor(game, chosenFactionId),
-        relation: getRelation(game, _humanId, chosenFactionId),
+        factionDisplayName: diplomacyDetailDisplayNameFor(
+          game,
+          chosenFactionId,
+        ),
+        kind: diplomacyDetailKindFor(game, chosenFactionId),
+        relation: getRelation(game, diplomacyDetailHumanId, chosenFactionId),
       );
       expect(find.text('DIPLOMATIC HISTORY'), findsOneWidget);
       if (history.isEmpty) {
@@ -187,7 +67,9 @@ void main() {
         );
       } else {
         expect(
-          find.text(formatDiplomaticEvent(history.first, game, _humanId)),
+          find.text(
+            formatDiplomaticEvent(history.first, game, diplomacyDetailHumanId),
+          ),
           findsOneWidget,
         );
       }
@@ -197,20 +79,24 @@ void main() {
   testWidgets(
     'DiplomacyDetailScreen hides Dossier when kind != greatPower and relation is null (empty history)',
     (WidgetTester tester) async {
-      final game = _minimalGame();
-      final allFactionIds = _otherFactionIds(game);
+      final game = diplomacyDetailMinimalGame();
+      final allFactionIds = diplomacyDetailOtherFactionIds(game);
       var factionId = allFactionIds.first;
       for (final id in allFactionIds) {
-        if (diplomaticHistoryForPair(game, _humanId, id).isEmpty) {
+        if (diplomaticHistoryForPair(
+          game,
+          diplomacyDetailHumanId,
+          id,
+        ).isEmpty) {
           factionId = id;
           break;
         }
       }
-      await _pumpDetail(
+      await pumpDiplomacyDetail(
         tester,
         game: game,
         factionId: factionId,
-        factionDisplayName: _displayNameFor(game, factionId),
+        factionDisplayName: diplomacyDetailDisplayNameFor(game, factionId),
         kind: FactionKind.minor,
         relation: null,
       );
@@ -224,103 +110,118 @@ void main() {
   );
 
   group('history / dossier matrix (GAME30002)', () {
-    for (final c in <({
-      String name,
-      Game Function() game,
-      FactionKind kind,
-      bool useRelation,
-      List<_Pin> pins,
-    })>[
-      (
-        name: 'GP shows dossier header and empty evidence',
-        game: _minimalGame,
-        kind: FactionKind.greatPower,
-        useRelation: true,
-        pins: [
-          (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
-          (find.text('DOSSIER'), findsOneWidget),
-          (find.textContaining('No dossier evidence yet.'), findsOneWidget),
-        ],
-      ),
-      (
-        name: 'GP empty history when relation is null',
-        game: _minimalGame,
-        kind: FactionKind.greatPower,
-        useRelation: false,
-        pins: [
-          (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
-          (find.text('DOSSIER'), findsOneWidget),
-          (find.text('No dossier evidence yet.'), findsOneWidget),
-          (find.text('No recorded events with this faction.'), findsOneWidget),
-        ],
-      ),
-      (
-        name: 'non-GP with relation shows Peace and hides Dossier',
-        game: _minimalGame,
-        kind: FactionKind.minor,
-        useRelation: true,
-        pins: [
-          (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
-          (find.text('DOSSIER'), findsNothing),
-          (find.text('No recorded events with this faction.'), findsOneWidget),
-          (find.textContaining('Peace'), findsOneWidget),
-        ],
-      ),
-      (
-        name: 'GP empty-state dossier when no evidence exists',
-        game: () => _minimalGame(eventType: DiplomaticEventType.declareWar),
-        kind: FactionKind.greatPower,
-        useRelation: true,
-        pins: [
-          (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
-          (find.text('DOSSIER'), findsOneWidget),
-          (find.text('No recorded events with this faction.'), findsOneWidget),
-          (find.text('No dossier evidence yet.'), findsOneWidget),
-          (find.textContaining('Peace'), findsOneWidget),
-        ],
-      ),
-      (
-        name: 'GP non-empty history + dossier at war',
-        game: () => _minimalGame(
-          eventType: DiplomaticEventType.declareWar,
-          includeHistory: true,
-          includeDossier: true,
-          atWar: true,
-        ),
-        kind: FactionKind.greatPower,
-        useRelation: true,
-        pins: [
-          (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
-          (find.text('DOSSIER'), findsOneWidget),
-          (find.textContaining('Turn 3:'), findsOneWidget),
-          (find.textContaining('evidence-1'), findsOneWidget),
-          (find.textContaining('declared war'), findsOneWidget),
-          (find.textContaining('War'), findsOneWidget),
-        ],
-      ),
-      (
-        name: 'GP non-empty history with empty Dossier at war',
-        game: () => _minimalGame(
-          eventType: DiplomaticEventType.declareWar,
-          includeHistory: true,
-          atWar: true,
-        ),
-        kind: FactionKind.greatPower,
-        useRelation: true,
-        pins: [
-          (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
-          (find.text('DOSSIER'), findsOneWidget),
-          (find.textContaining('declared war'), findsOneWidget),
-          (find.textContaining('War'), findsOneWidget),
-          (find.text('No dossier evidence yet.'), findsOneWidget),
-        ],
-      ),
-    ]) {
+    for (final c
+        in <
+          ({
+            String name,
+            Game Function() game,
+            FactionKind kind,
+            bool useRelation,
+            List<DiplomacyDetailPin> pins,
+          })
+        >[
+          (
+            name: 'GP shows dossier header and empty evidence',
+            game: diplomacyDetailMinimalGame,
+            kind: FactionKind.greatPower,
+            useRelation: true,
+            pins: [
+              (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
+              (find.text('DOSSIER'), findsOneWidget),
+              (find.textContaining('No dossier evidence yet.'), findsOneWidget),
+            ],
+          ),
+          (
+            name: 'GP empty history when relation is null',
+            game: diplomacyDetailMinimalGame,
+            kind: FactionKind.greatPower,
+            useRelation: false,
+            pins: [
+              (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
+              (find.text('DOSSIER'), findsOneWidget),
+              (find.text('No dossier evidence yet.'), findsOneWidget),
+              (
+                find.text('No recorded events with this faction.'),
+                findsOneWidget,
+              ),
+            ],
+          ),
+          (
+            name: 'non-GP with relation shows Peace and hides Dossier',
+            game: diplomacyDetailMinimalGame,
+            kind: FactionKind.minor,
+            useRelation: true,
+            pins: [
+              (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
+              (find.text('DOSSIER'), findsNothing),
+              (
+                find.text('No recorded events with this faction.'),
+                findsOneWidget,
+              ),
+              (find.textContaining('Peace'), findsOneWidget),
+            ],
+          ),
+          (
+            name: 'GP empty-state dossier when no evidence exists',
+            game: () => diplomacyDetailMinimalGame(
+              eventType: DiplomaticEventType.declareWar,
+            ),
+            kind: FactionKind.greatPower,
+            useRelation: true,
+            pins: [
+              (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
+              (find.text('DOSSIER'), findsOneWidget),
+              (
+                find.text('No recorded events with this faction.'),
+                findsOneWidget,
+              ),
+              (find.text('No dossier evidence yet.'), findsOneWidget),
+              (find.textContaining('Peace'), findsOneWidget),
+            ],
+          ),
+          (
+            name: 'GP non-empty history + dossier at war',
+            game: () => diplomacyDetailMinimalGame(
+              eventType: DiplomaticEventType.declareWar,
+              includeHistory: true,
+              includeDossier: true,
+              atWar: true,
+            ),
+            kind: FactionKind.greatPower,
+            useRelation: true,
+            pins: [
+              (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
+              (find.text('DOSSIER'), findsOneWidget),
+              (find.textContaining('Turn 3:'), findsOneWidget),
+              (find.textContaining('evidence-1'), findsOneWidget),
+              (find.textContaining('declared war'), findsOneWidget),
+              (find.textContaining('War'), findsOneWidget),
+            ],
+          ),
+          (
+            name: 'GP non-empty history with empty Dossier at war',
+            game: () => diplomacyDetailMinimalGame(
+              eventType: DiplomaticEventType.declareWar,
+              includeHistory: true,
+              atWar: true,
+            ),
+            kind: FactionKind.greatPower,
+            useRelation: true,
+            pins: [
+              (find.text('DIPLOMATIC HISTORY'), findsOneWidget),
+              (find.text('DOSSIER'), findsOneWidget),
+              (find.textContaining('declared war'), findsOneWidget),
+              (find.textContaining('War'), findsOneWidget),
+              (find.text('No dossier evidence yet.'), findsOneWidget),
+            ],
+          ),
+        ]) {
       testWidgets(c.name, (WidgetTester tester) async {
         final game = c.game();
-        final relation =
-            c.useRelation ? getRelation(game, _humanId, _otherId) : null;
-        await _pumpOtherGp(
+        final relation = c.useRelation
+            ? getRelation(game, diplomacyDetailHumanId, diplomacyDetailOtherId)
+            : null;
+        await pumpDiplomacyDetailOtherGp(
           tester,
           game: game,
           kind: c.kind,
@@ -336,40 +237,10 @@ void main() {
   testWidgets(
     'DiplomacyDetailScreen falls back to "Unknown faction" for events with unknown relation pairs',
     (WidgetTester tester) async {
-      const unknownFactionId = 'minorX';
-      final game = Game(
-        id: 'test-unknown',
-        worldState: WorldState(
-          turnState: TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: const RegionData(),
-          newWorld: const RegionData(),
-        ),
-        turnTimeMapping: TurnTimeMapping.gdd01,
-        players: [
-          Player(
-            id: _humanId,
-            displayName: 'Human GP',
-            isHuman: true,
-            treasury: 0,
-          ),
-        ],
-        diplomacyRelations: const [],
-        diplomaticHistoryEvents: [
-          DiplomaticEvent(
-            turn: 2,
-            intraTurnIndex: 0,
-            type: DiplomaticEventType.declareWar,
-            participants: {_humanId, unknownFactionId},
-            fromFactionId: _humanId,
-            toFactionId: unknownFactionId,
-          ),
-        ],
-        dossierEvidenceEntries: const [],
-      );
-      await _pumpDetail(
+      await pumpDiplomacyDetail(
         tester,
-        game: game,
-        factionId: unknownFactionId,
+        game: diplomacyDetailUnknownFactionGame(),
+        factionId: diplomacyDetailUnknownFactionId,
         factionDisplayName: 'Unknown Faction',
         kind: FactionKind.minor,
         relation: null,
@@ -380,45 +251,52 @@ void main() {
     },
   );
 
-  testWidgets(
-    'DiplomacyDetailScreen renders dark editorial-monocle chrome '
-    '(CtTopBar + scaffold bg) per Refs #2863 S5',
-    (WidgetTester tester) async {
-      final game = _minimalGame();
-      await _pumpOtherGp(
-        tester,
-        game: game,
-        relation: getRelation(game, _humanId, _otherId),
-      );
-      expect(find.byType(CtTopBar), findsOneWidget);
-      expect(find.byType(CtBackButton), findsOneWidget);
-      expect(find.byType(AppBar), findsNothing);
-      expect(find.byType(CtGameFeatureScreenShell), findsOneWidget);
-      final CtGameFeatureScreenShell shell = tester.widget(
-        find.byType(CtGameFeatureScreenShell),
-      );
-      expect(shell.backgroundColor, EditorialMonoclePalette.bg);
-      expect(shell.attachGameToUiListener, isFalse);
-      final Scaffold scaffold = tester.widget(find.byType(Scaffold));
-      expect(scaffold.backgroundColor, EditorialMonoclePalette.bg);
-    },
-  );
+  testWidgets('DiplomacyDetailScreen renders dark editorial-monocle chrome '
+      '(CtTopBar + scaffold bg) per Refs #2863 S5', (
+    WidgetTester tester,
+  ) async {
+    final game = diplomacyDetailMinimalGame();
+    await pumpDiplomacyDetailOtherGp(
+      tester,
+      game: game,
+      relation: getRelation(
+        game,
+        diplomacyDetailHumanId,
+        diplomacyDetailOtherId,
+      ),
+    );
+    expect(find.byType(CtTopBar), findsOneWidget);
+    expect(find.byType(CtBackButton), findsOneWidget);
+    expect(find.byType(AppBar), findsNothing);
+    expect(find.byType(CtGameFeatureScreenShell), findsOneWidget);
+    final CtGameFeatureScreenShell shell = tester.widget(
+      find.byType(CtGameFeatureScreenShell),
+    );
+    expect(shell.backgroundColor, EditorialMonoclePalette.bg);
+    expect(shell.attachGameToUiListener, isFalse);
+    final Scaffold scaffold = tester.widget(find.byType(Scaffold));
+    expect(scaffold.backgroundColor, EditorialMonoclePalette.bg);
+  });
 
   testWidgets(
     'DiplomacyDetailScreen emits exactly one PopNavigationEvent when the '
     'CtTopBar back button is tapped per Refs #2863 S5',
     (WidgetTester tester) async {
-      final game = _minimalGame();
+      final game = diplomacyDetailMinimalGame();
       final AppEventBus bus = AppEventBus.create();
       addTearDown(bus.dispose);
       final List<PopNavigationEvent> popEvents = <PopNavigationEvent>[];
       final sub = bus.on<PopNavigationEvent>().listen(popEvents.add);
       addTearDown(sub.cancel);
-      await _pumpOtherGp(
+      await pumpDiplomacyDetailOtherGp(
         tester,
         overrides: [appEventBusProvider.overrideWith((ref) => bus)],
         game: game,
-        relation: getRelation(game, _humanId, _otherId),
+        relation: getRelation(
+          game,
+          diplomacyDetailHumanId,
+          diplomacyDetailOtherId,
+        ),
       );
       expect(popEvents, isEmpty);
       await tester.tap(find.byType(CtBackButton));
@@ -428,64 +306,76 @@ void main() {
   );
 
   group('current relation card chrome', () {
-    for (final c in <({
-      String name,
-      Game Function() game,
-      void Function(WidgetTester tester) assertUi,
-    })>[
-      (
-        name: 'War label in --danger colour',
-        game: () => _minimalGame(
-          eventType: DiplomaticEventType.declareWar,
-          atWar: true,
-        ),
-        assertUi: (tester) {
-          expect(find.text('CURRENT RELATION'), findsOneWidget);
-          final Text war = tester.widget(find.text('War'));
-          expect(war.style?.color, EditorialMonoclePalette.danger);
-        },
-      ),
-      (
-        name: 'Peace label in --success colour',
-        game: _minimalGame,
-        assertUi: (tester) {
-          expect(find.text('CURRENT RELATION'), findsOneWidget);
-          final Text peace = tester.widget(find.text('Peace'));
-          expect(peace.style?.color, EditorialMonoclePalette.success);
-        },
-      ),
-      (
-        name: 'ALLIANCE badge in --accent for formal alliance (Refs #3625 AC4)',
-        game: () => _minimalGame(
-          eventType: DiplomaticEventType.allianceFormed,
-          score: 90,
-          formalAlliance: true,
-        ),
-        assertUi: (tester) {
-          expect(find.text('CURRENT RELATION'), findsOneWidget);
-          final Finder badge = find.text(kDiplomacyAllianceBadgeLabel);
-          expect(badge, findsOneWidget);
-          final Text badgeText = tester.widget<Text>(badge);
-          expect(badgeText.style?.color, EditorialMonoclePalette.accent);
-          expect(kDiplomacyAllianceBadgeLabel, isNot('Friendly'));
-        },
-      ),
-      (
-        name:
-            'omits ALLIANCE badge for informal Allied band (Refs #3625 AC4 negative)',
-        game: () => _minimalGame(score: 90),
-        assertUi: (tester) {
-          expect(find.text('CURRENT RELATION'), findsOneWidget);
-          expect(find.text(kDiplomacyAllianceBadgeLabel), findsNothing);
-          expect(find.textContaining('Devoted'), findsOneWidget);
-        },
-      ),
-    ]) {
+    for (final c
+        in <
+          ({
+            String name,
+            Game Function() game,
+            void Function(WidgetTester tester) assertUi,
+          })
+        >[
+          (
+            name: 'War label in --danger colour',
+            game: () => diplomacyDetailMinimalGame(
+              eventType: DiplomaticEventType.declareWar,
+              atWar: true,
+            ),
+            assertUi: (tester) {
+              expect(find.text('CURRENT RELATION'), findsOneWidget);
+              final Text war = tester.widget(find.text('War'));
+              expect(war.style?.color, EditorialMonoclePalette.danger);
+            },
+          ),
+          (
+            name: 'Peace label in --success colour',
+            game: diplomacyDetailMinimalGame,
+            assertUi: (tester) {
+              expect(find.text('CURRENT RELATION'), findsOneWidget);
+              final Text peace = tester.widget(find.text('Peace'));
+              expect(peace.style?.color, EditorialMonoclePalette.success);
+            },
+          ),
+          (
+            name:
+                'ALLIANCE badge in --accent for formal alliance (Refs #3625 AC4)',
+            game: () => diplomacyDetailMinimalGame(
+              eventType: DiplomaticEventType.allianceFormed,
+              score: 90,
+              formalAlliance: true,
+            ),
+            assertUi: (tester) {
+              expect(find.text('CURRENT RELATION'), findsOneWidget);
+              final Finder badge = find.text(kDiplomacyAllianceBadgeLabel);
+              expect(badge, findsOneWidget);
+              final Text badgeText = tester.widget<Text>(badge);
+              expect(badgeText.style?.color, EditorialMonoclePalette.accent);
+              expect(kDiplomacyAllianceBadgeLabel, isNot('Friendly'));
+            },
+          ),
+          (
+            name:
+                'omits ALLIANCE badge for informal Allied band (Refs #3625 AC4 negative)',
+            game: () => diplomacyDetailMinimalGame(score: 90),
+            assertUi: (tester) {
+              expect(find.text('CURRENT RELATION'), findsOneWidget);
+              expect(find.text(kDiplomacyAllianceBadgeLabel), findsNothing);
+              expect(find.textContaining('Devoted'), findsOneWidget);
+            },
+          ),
+        ]) {
       testWidgets(c.name, (WidgetTester tester) async {
         final game = c.game();
-        final relation = getRelation(game, _humanId, _otherId);
+        final relation = getRelation(
+          game,
+          diplomacyDetailHumanId,
+          diplomacyDetailOtherId,
+        );
         expect(relation, isNotNull);
-        await _pumpOtherGp(tester, game: game, relation: relation);
+        await pumpDiplomacyDetailOtherGp(
+          tester,
+          game: game,
+          relation: relation,
+        );
         c.assertUi(tester);
       });
     }
