+++ app/test/overture_dialogue_overlay_test.dart
@@ -10,69 +10,11 @@ import 'package:colonizethis_app/features/game/widgets/dialogue/overture_dialogu
 import 'package:colonizethis_app/widgets/ct_nine_patch_button.dart';
 import 'package:colonizethis_app_ui_chrome/widgets/ct_brass_divider.dart';
 import 'package:colonizethis_app/widgets/ct_dialog_shell.dart';
-import 'package:colonizethis_app/widgets/ct_toggle_switch.dart';
 
-import 'app_shell_harness.dart';
-
-const OvertureOffer _gp2TradeConsulate = OvertureOffer(
-  offererGpId: 'gp2',
-  targetFactionId: 'gp1',
-  stage: OvertureStage.tradeConsulate,
-);
-
-const OvertureOffer _gp2Embassy = OvertureOffer(
-  offererGpId: 'gp2',
-  targetFactionId: 'gp1',
-  stage: OvertureStage.embassy,
-);
-
-const List<OvertureOffer> _singleGp2Offer = [_gp2TradeConsulate];
-const List<OvertureOffer> _twoStageGp2Offers = [
-  _gp2TradeConsulate,
-  _gp2Embassy,
-];
+import 'overture_dialogue_overlay_test_support.dart';
 
 void main() {
   suppressLogsForTests();
-
-  Game game() {
-    return const Game(
-      id: 'g1',
-      worldState: WorldState(
-        turnState: TurnState(phase: TurnPhase.orders, turnNumber: 0),
-        oldWorld: RegionData(),
-        newWorld: RegionData(),
-      ),
-      players: [
-        Player(id: 'gp1', displayName: 'Great Power 1', isHuman: true),
-        Player(id: 'gp2', displayName: 'Great Power 2', isHuman: false),
-      ],
-    );
-  }
-
-  Future<void> pumpOverlay(
-    WidgetTester tester, {
-    List<OvertureOffer> offers = _singleGp2Offer,
-    void Function(List<OvertureDecision>)? onDecisions,
-    Size surfaceSize = const Size(900, 900),
-  }) async {
-    addTearDown(tester.view.reset);
-    tester.view.physicalSize = surfaceSize;
-    tester.view.devicePixelRatio = 1.0;
-    // Editorial shell via buildAppShell (Refs #4035 — no inline MaterialApp).
-    await tester.pumpWidget(
-      buildAppShell(
-        child: OvertureDialogueOverlay(
-          game: game(),
-          pendingOvertures: offers,
-          skipIntroForTest: true,
-          onDecisions: onDecisions ?? (_) {},
-          child: const Scaffold(body: Text('child')),
-        ),
-      ),
-    );
-    await tester.pump();
-  }
 
   group('OvertureDialogueOverlay', () {
     testWidgets(
@@ -80,9 +22,9 @@ void main() {
       (WidgetTester tester) async {
         List<OvertureDecision>? submitted;
 
-        await pumpOverlay(
+        await pumpOvertureOverlay(
           tester,
-          offers: _twoStageGp2Offers,
+          offers: twoStageGp2OvertureOffers,
           onDecisions: (d) => submitted = List.of(d),
         );
 
@@ -123,7 +65,7 @@ void main() {
       'phase 2 Submit is disabled until every row has a non-null decision '
       '(#2867 R23 / AC4 — positive enable transition)',
       (WidgetTester tester) async {
-        await pumpOverlay(tester, offers: _twoStageGp2Offers);
+        await pumpOvertureOverlay(tester, offers: twoStageGp2OvertureOffers);
 
         final Finder submitFinder = find.byKey(
           const ValueKey<String>('overtureSubmitButton'),
@@ -163,46 +105,44 @@ void main() {
       },
     );
 
-    testWidgets(
-      'phase 2 Submit disabled while only the second row is decided '
-      '(#2867 R23 / AC4 — negative case)',
-      (WidgetTester tester) async {
-        List<OvertureDecision>? submitted;
+    testWidgets('phase 2 Submit disabled while only the second row is decided '
+        '(#2867 R23 / AC4 — negative case)', (WidgetTester tester) async {
+      List<OvertureDecision>? submitted;
 
-        await pumpOverlay(
-          tester,
-          offers: _twoStageGp2Offers,
-          onDecisions: (d) => submitted = List.of(d),
-        );
+      await pumpOvertureOverlay(
+        tester,
+        offers: twoStageGp2OvertureOffers,
+        onDecisions: (d) => submitted = List.of(d),
+      );
 
-        await tester.tap(find.text('Reject').last);
-        await tester.pump();
+      await tester.tap(find.text('Reject').last);
+      await tester.pump();
 
-        final Finder submitFinder = find.byKey(
-          const ValueKey<String>('overtureSubmitButton'),
-        );
-        final CtNinePatchButton submitButton = tester
-            .widget<CtNinePatchButton>(submitFinder);
-        expect(submitButton.enabled, isFalse);
+      final Finder submitFinder = find.byKey(
+        const ValueKey<String>('overtureSubmitButton'),
+      );
+      final CtNinePatchButton submitButton = tester.widget<CtNinePatchButton>(
+        submitFinder,
+      );
+      expect(submitButton.enabled, isFalse);
 
-        // `warnIfMissed: false` because the disabled button intentionally
-        // ignores hit-tests (per CtNinePatchButton § Disabled).
-        await tester.tap(find.text('Submit'), warnIfMissed: false);
-        await tester.pump();
-        expect(
-          submitted,
-          isNull,
-          reason:
-              'Tapping the disabled Submit must not invoke onDecisions '
-              'while any row is still undecided (#2867 R23).',
-        );
-      },
-    );
+      // `warnIfMissed: false` because the disabled button intentionally
+      // ignores hit-tests (per CtNinePatchButton § Disabled).
+      await tester.tap(find.text('Submit'), warnIfMissed: false);
+      await tester.pump();
+      expect(
+        submitted,
+        isNull,
+        reason:
+            'Tapping the disabled Submit must not invoke onDecisions '
+            'while any row is still undecided (#2867 R23).',
+      );
+    });
 
     testWidgets(
       'phase 2 title uses --accent color and 0.05em letter-spacing (#2867 R2/R21)',
       (WidgetTester tester) async {
-        await pumpOverlay(tester);
+        await pumpOvertureOverlay(tester);
 
         final Finder titleFinder = find.byKey(
           const ValueKey<String>('overtureTitle'),
@@ -223,23 +163,20 @@ void main() {
     testWidgets(
       'phase 2 renders CtBrassDivider between title and intro (#2867 R21)',
       (WidgetTester tester) async {
-        await pumpOverlay(tester);
+        await pumpOvertureOverlay(tester);
 
         final Finder dividerFinder = find.byKey(
           const ValueKey<String>('overtureBrassDivider'),
         );
         expect(dividerFinder, findsOneWidget);
-        expect(
-          dividerFinder.evaluate().single.widget,
-          isA<CtBrassDivider>(),
-        );
+        expect(dividerFinder.evaluate().single.widget, isA<CtBrassDivider>());
       },
     );
 
     testWidgets(
       'phase 2 intro is rendered in --muted italic body style (#2867 R5/R21)',
       (WidgetTester tester) async {
-        await pumpOverlay(tester);
+        await pumpOvertureOverlay(tester);
 
         final Finder introFinder = find.byKey(
           const ValueKey<String>('overtureIntro'),
@@ -254,7 +191,7 @@ void main() {
     testWidgets(
       'offer row paints offerer in --accent and stage in --muted (#2867 R22)',
       (WidgetTester tester) async {
-        await pumpOverlay(tester);
+        await pumpOvertureOverlay(tester);
 
         final Finder offererFinder = find.byKey(
           const ValueKey<String>('overtureOfferOfferer'),
@@ -285,7 +222,7 @@ void main() {
     testWidgets(
       'phase 2 chrome contains no Material AlertDialog/ListTile/Card chrome (#2867 R1)',
       (WidgetTester tester) async {
-        await pumpOverlay(tester);
+        await pumpOvertureOverlay(tester);
 
         final Finder overlay = find.byType(OvertureDialogueOverlay);
         expect(
@@ -303,173 +240,40 @@ void main() {
       },
     );
 
-    testWidgets(
-      'phase 2 scrim resolves to EditorialMonoclePalette.dialogScrim '
-      '(#2867 R1; mirrors intervention overlay S9)',
-      (WidgetTester tester) async {
-        await pumpOverlay(tester);
+    testWidgets('phase 2 scrim resolves to EditorialMonoclePalette.dialogScrim '
+        '(#2867 R1; mirrors intervention overlay S9)', (
+      WidgetTester tester,
+    ) async {
+      await pumpOvertureOverlay(tester);
 
-        final Finder shellFinder = find.byType(
-          CtDialogShell,
-        );
-        expect(shellFinder, findsOneWidget);
-        final Material scrim = tester.widget<Material>(
-          find
-              .ancestor(of: shellFinder, matching: find.byType(Material))
-              .first,
-        );
-        expect(scrim.color, EditorialMonoclePalette.dialogScrim);
-        expect(scrim.color, isNot(Colors.black54));
-      },
-    );
+      final Finder shellFinder = find.byType(CtDialogShell);
+      expect(shellFinder, findsOneWidget);
+      final Material scrim = tester.widget<Material>(
+        find.ancestor(of: shellFinder, matching: find.byType(Material)).first,
+      );
+      expect(scrim.color, EditorialMonoclePalette.dialogScrim);
+      expect(scrim.color, isNot(Colors.black54));
+    });
 
-    testWidgets(
-      'no Material descendant uses the legacy Colors.black54 scrim '
-      '(#2867 R1 negative regression guard)',
-      (WidgetTester tester) async {
-        await pumpOverlay(tester);
+    testWidgets('no Material descendant uses the legacy Colors.black54 scrim '
+        '(#2867 R1 negative regression guard)', (WidgetTester tester) async {
+      await pumpOvertureOverlay(tester);
 
-        final Finder overlay = find.byType(OvertureDialogueOverlay);
-        for (final Element element in find
-            .descendant(of: overlay, matching: find.byType(Material))
-            .evaluate()) {
-          final Material material = element.widget as Material;
-          expect(
-            material.color,
-            isNot(Colors.black54),
-            reason:
-                'Legacy Colors.black54 scrim must not leak into the overture '
-                'overlay; use EditorialMonoclePalette.dialogScrim per '
-                '#2867 R1 / SPEC/ui/pixel-art-ui-catalog.md § Dialog scrim.',
-          );
-        }
-      },
-    );
-  });
-
-  group('OvertureDialogueOverlay R22 CtToggleSwitch (#2867 R22)', () {
-    Finder acceptToggleAt(int rowIndex) => find.byKey(
-      ValueKey<String>('overtureAcceptToggle_$rowIndex'),
-    );
-
-    Finder rejectToggleAt(int rowIndex) => find.byKey(
-      ValueKey<String>('overtureRejectToggle_$rowIndex'),
-    );
-
-    testWidgets(
-      'phase 2: Accept/Reject CtToggleSwitch glow tokens; only Submit is nine-patch',
-      (WidgetTester tester) async {
-        await pumpOverlay(tester, offers: _twoStageGp2Offers);
-
-        for (var i = 0; i < 2; i++) {
-          expect(acceptToggleAt(i), findsOneWidget);
-          expect(rejectToggleAt(i), findsOneWidget);
-
-          final CtToggleSwitch accept = tester.widget<CtToggleSwitch>(
-            acceptToggleAt(i),
-          );
-          expect(accept.value, isFalse);
-          expect(accept.onGlowColor, EditorialMonoclePalette.success);
-
-          final CtToggleSwitch reject = tester.widget<CtToggleSwitch>(
-            rejectToggleAt(i),
-          );
-          expect(reject.value, isFalse);
-          expect(reject.onGlowColor, EditorialMonoclePalette.danger);
-        }
-
-        // Submit is the only CtNinePatchButton in phase 2 (Accept/Reject are
-        // CtToggleSwitch per #2867 R22).
-        final Iterable<CtNinePatchButton> buttons = tester
-            .widgetList<CtNinePatchButton>(find.byType(CtNinePatchButton));
-        expect(buttons, hasLength(1));
+      final Finder overlay = find.byType(OvertureDialogueOverlay);
+      for (final Element element
+          in find
+              .descendant(of: overlay, matching: find.byType(Material))
+              .evaluate()) {
+        final Material material = element.widget as Material;
         expect(
-          buttons.single.key,
-          const ValueKey<String>('overtureSubmitButton'),
-        );
-      },
-    );
-
-    testWidgets(
-      'Accept/Reject toggles are mutually exclusive (on turns the other off)',
-      (WidgetTester tester) async {
-        await pumpOverlay(tester);
-
-        await tester.tap(acceptToggleAt(0));
-        await tester.pump();
-        expect(tester.widget<CtToggleSwitch>(acceptToggleAt(0)).value, isTrue);
-        expect(tester.widget<CtToggleSwitch>(rejectToggleAt(0)).value, isFalse);
-
-        await tester.tap(rejectToggleAt(0));
-        await tester.pump();
-        expect(tester.widget<CtToggleSwitch>(rejectToggleAt(0)).value, isTrue);
-        expect(tester.widget<CtToggleSwitch>(acceptToggleAt(0)).value, isFalse);
-      },
-    );
-
-    testWidgets(
-      'tapping Accept then Reject swaps the committed decision to false',
-      (WidgetTester tester) async {
-        List<OvertureDecision>? submitted;
-        await pumpOverlay(
-          tester,
-          offers: _singleGp2Offer,
-          onDecisions: (d) => submitted = List.of(d),
-        );
-
-        await tester.tap(acceptToggleAt(0));
-        await tester.pump();
-        await tester.tap(rejectToggleAt(0));
-        await tester.pump();
-        await tester.tap(find.byKey(const ValueKey<String>('overtureSubmitButton')));
-        await tester.pump();
-
-        expect(submitted, isNotNull);
-        expect(submitted, hasLength(1));
-        expect(submitted!.first.accepted, isFalse);
-      },
-    );
-
-    testWidgets(
-      'tapping a currently-on toggle reverts the row to undecided and '
-      're-engages the #2867 R23 Submit gate (positive R22 + R23 interaction)',
-      (WidgetTester tester) async {
-        await pumpOverlay(tester);
-
-        final Finder submitFinder = find.byKey(
-          const ValueKey<String>('overtureSubmitButton'),
-        );
-
-        await tester.tap(acceptToggleAt(0));
-        await tester.pump();
-        expect(
-          tester.widget<CtNinePatchButton>(submitFinder).enabled,
-          isTrue,
+          material.color,
+          isNot(Colors.black54),
           reason:
-              'Single-row overlay enables Submit immediately once a row is '
-              'decided (#2867 R23 positive case).',
+              'Legacy Colors.black54 scrim must not leak into the overture '
+              'overlay; use EditorialMonoclePalette.dialogScrim per '
+              '#2867 R1 / SPEC/ui/pixel-art-ui-catalog.md § Dialog scrim.',
         );
-
-        // Tap the Accept toggle again while currently on -> reverts to
-        // undecided and Submit must disable again.
-        await tester.tap(acceptToggleAt(0));
-        await tester.pump();
-        final CtToggleSwitch accept = tester.widget<CtToggleSwitch>(
-          acceptToggleAt(0),
-        );
-        final CtToggleSwitch reject = tester.widget<CtToggleSwitch>(
-          rejectToggleAt(0),
-        );
-        expect(accept.value, isFalse);
-        expect(reject.value, isFalse);
-        expect(
-          tester.widget<CtNinePatchButton>(submitFinder).enabled,
-          isFalse,
-          reason:
-              'Reverting a row to undecided must re-engage the R23 Submit '
-              'gate so the user cannot submit unintentional decisions.',
-        );
-      },
-    );
+      }
+    });
   });
 }
